remove warning and add two missing file needed
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@120381 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: a11641ee017fe967fd3305a4fe3a4c7341098118

+++ Rec/Rec/RecoUpgradeMC/src/CheatedVeloPixPat.cpp
@@ -124,7 +124,7 @@ StatusCode CheatedVeloPixPat::execute() {
     unsigned int prevSensor = 200;
     // Loop over the LHCbID corresponding to this MCParticle
 
-    if(part->key() >= linkedIds.size()) continue;
+    if( (unsigned int) part->key() >= linkedIds.size()) continue;
     for ( std::vector<int>::const_iterator itIm = linkedIds[part->key()].begin();
           linkedIds[part->key()].end() != itIm; ++itIm ) {
       // Get the Lite cluster corresponding to the LHCbID
